requirements.txt
fastapi
uvicorn
motor
pymongo
langchain
langchain-community
litellm
openai
langchain-openai
python-dotenv
llama-index
neo4j
pdfplumber
PyPDF2
nltk
llama-index
llama-index-core
pypdf
pymupdf
llama-index-vector-stores-redis
langfuse
llama-index-embeddings-azure-openai
llama-index-llms-azure-openai
asyncio